+++ app/models/role_certificate_requirement.rb
@@ -1,0 +1,6 @@
+class RoleCertificateRequirement < ApplicationRecord
+  belongs_to :role
+  belongs_to :certificate
+
+  validates :role_id, uniqueness: { scope: :certificate_id }
+end